refactor(bars): export quantizeBarsToLastClosed from barrel (barrel-only import)
