+++ src/entities/bars/index.ts
@@ -1,1 +1,2 @@
 export { getBarsStatic } from './lib/barsStaticCache';
+export { quantizeBarsToLastClosed } from './lib/quantizeBars';
